--- code_ops/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "code_ops"
-version = "0.1.0"
-edition = "2021"
-rust-version.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tree-sitter = "0.22.5"
-tree-sitter-rust = "0.21.0"
-tree-sitter-python = "0.21.0"
-tree-sitter-ruby = "0.21.0"
-tree-sitter-typescript = "0.21.1"
-tree-sitter-javascript = "0.21.3"
-strum = { version = "0.26.2", features = ["derive"] }
-strum_macros = "0.26.2"
-
-derive_builder = { workspace = true }
-anyhow = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-
-[dev-dependencies]
-indoc = { workspace = true }
-
-[build-dependencies]
-cc = "1.0.95"
