+++ indexing/Cargo.toml
@@ -5,27 +5,35 @@ edition = "2021"
 rust-version.workspace = true
 
 [dependencies]
-ignore = "0.4.22"
-code_ops = { path = "../code_ops" }
-qdrant-client = { version = "1.9.0" }
+anyhow = { version = "1.0.86", features = ["backtrace"] }
 async-stream = "0.3.5"
-async-openai = { workspace = true }
+async-trait = "0.1.80"
+async-openai = "0.23.2"
+chrono = { version = "0.4.38" }
+derive_builder = "0.20.0"
 futures-util = "0.3.30"
-# futures-util = { workspace = true }
+ignore = "0.4.22"
+indoc = "2.0.5"
+itertools = { version = "0.13.0" }
+serde = { version = "1.0.203", features = ["derive"] }
+serde_json = "1.0.117"
+text-splitter = { version = "0.13.1", features = ["markdown"] }
+tokio = { version = "1.38.0", features = ["full"] }
+tracing = { version = "0.1.40", features = ["log"] }
+strum = "0.26.2"
+strum_macros = "0.26.4"
+
+# Integrations
+qdrant-client = { version = "1.9.0" }
 redis = { version = "0.25.4", features = [
   "aio",
   "tokio-comp",
   "connection-manager",
   "tokio-rustls-comp",
 ] }
-text-splitter = { version = "0.13.1", features = ["markdown"] }
-
-chrono = { workspace = true }
-anyhow = { workspace = true }
-tracing = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-itertools = { workspace = true }
-indoc = { workspace = true }
-async-trait = { workspace = true }
-tokio = { workspace = true }
+tree-sitter = "0.22.5"
+tree-sitter-rust = "0.21.0"
+tree-sitter-python = "0.21.0"
+tree-sitter-ruby = "0.21.0"
+tree-sitter-typescript = "0.21.1"
+tree-sitter-javascript = "0.21.3"
