+++ indexing/src/embeddings.rs
@@ -1,3 +1,1 @@
-pub mod openai;
-
 pub type Embeddings = Vec<Vec<f32>>;
