+++ indexing/src/ingestion/ingestion_node.rs
@@ -1,13 +1,7 @@
-use anyhow::{Context as _, Result};
 use std::{
     collections::HashMap,
     hash::{Hash, Hasher},
     path::PathBuf,
-};
-
-use qdrant_client::{
-    client::Payload,
-    qdrant::{self, Value},
 };
 
 #[derive(Debug, Default, Clone)]
@@ -43,36 +37,5 @@ impl Hash for IngestionNode {
     fn hash<H: Hasher>(&self, state: &mut H) {
         self.path.hash(state);
         self.chunk.hash(state);
-    }
-}
-
-impl TryInto<qdrant::PointStruct> for IngestionNode {
-    type Error = anyhow::Error;
-
-    fn try_into(mut self) -> Result<qdrant::PointStruct> {
-        let id = self.calculate_hash();
-
-        self.metadata.extend([
-            ("path".to_string(), self.path.to_string_lossy().to_string()),
-            ("content".to_string(), self.chunk),
-            (
-                "last_updated_at".to_string(),
-                chrono::Utc::now().to_rfc3339(),
-            ),
-        ]);
-
-        // Damn who build this api
-        let payload: Payload = self
-            .metadata
-            .iter()
-            .map(|(k, v)| (k.as_str(), Value::from(v.as_str())))
-            .collect::<HashMap<&str, Value>>()
-            .into();
-
-        Ok(qdrant::PointStruct::new(
-            id,
-            self.vector.context("Vector is not set")?,
-            payload,
-        ))
     }
 }
