+++ indexing/src/integrations/mod.rs
@@ -1,0 +1,4 @@
+pub mod openai;
+pub mod qdrant;
+pub mod redis;
+pub mod treesitter;
