+++ indexing/src/integrations/redis/mod.rs
@@ -1,0 +1,3 @@
+mod node_cache;
+
+pub use node_cache::RedisNodeCache;
