+++ indexing/src/integrations/treesitter/mod.rs
@@ -1,0 +1,5 @@
+mod splitter;
+mod supported_languages;
+
+pub use splitter::{ChunkSize, CodeSplitter, CodeSplitterBuilder};
+pub use supported_languages::SupportedLanguages;
