+++ indexing/src/lib.rs
@@ -1,12 +1,10 @@
-pub mod embedders;
+pub mod embeddings;
+pub mod ingestion;
+pub mod integrations;
 pub mod loaders;
-pub mod node_caches;
 pub mod query;
-pub mod storage;
+pub mod traits;
 pub mod transformers;
 
-mod ingestion_node;
-mod traits;
-
-mod ingestion_pipeline;
-pub use ingestion_pipeline::IngestionPipeline;
+pub use embeddings::*;
+pub use traits::*;
