--- indexing/src/node_caches/mod.rs
@@ -1,2 +1,0 @@
-pub mod redis;
-pub use redis::Redis;
