--- indexing/src/storage/mod.rs
@@ -1,3 +1,0 @@
-pub mod qdrant;
-
-pub use qdrant::Qdrant;
