+++ indexing/src/traits.rs
@@ -1,6 +1,6 @@
 use std::fmt::Debug;
 
-use crate::{ingestion_node::IngestionNode, ingestion_pipeline::IngestionStream};
+use crate::{ingestion::IngestionNode, ingestion::IngestionStream, Embeddings};
 use anyhow::Result;
 use async_trait::async_trait;
 
@@ -31,17 +31,6 @@ pub trait ChunkerTransformer: Send + Sync + Debug {
 }
 
 #[async_trait]
-/// Persists nodes
-pub trait Storage: Send + Sync {
-    async fn setup(&self) -> Result<()>;
-    async fn store(&self, node: IngestionNode) -> Result<()>;
-    async fn batch_store(&self, nodes: Vec<IngestionNode>) -> Result<()>;
-    fn batch_size(&self) -> Option<usize> {
-        None
-    }
-}
-
-#[async_trait]
 /// Caches nodes, typically by their path and hash
 /// Recommended to namespace on the storage
 ///
@@ -53,11 +42,22 @@ pub trait NodeCache: Send + Sync + Debug {
 
 #[async_trait]
 pub trait Embed: Debug + Send + Sync {
-    async fn embed(&self, input: Vec<String>) -> Result<crate::embedders::Embeddings>;
+    async fn embed(&self, input: Vec<String>) -> Result<Embeddings>;
 }
 
 #[async_trait]
 pub trait SimplePrompt: Debug + Send + Sync {
     // Takes a simple prompt, prompts the llm and returns the response
     async fn prompt(&self, prompt: &str) -> Result<String>;
+}
+
+#[async_trait]
+/// Persists nodes
+pub trait Storage: Send + Sync {
+    async fn setup(&self) -> Result<()>;
+    async fn store(&self, node: IngestionNode) -> Result<()>;
+    async fn batch_store(&self, nodes: Vec<IngestionNode>) -> Result<()>;
+    fn batch_size(&self) -> Option<usize> {
+        None
+    }
 }
